requirements.txt
# Django and Core Dependencies
Django==4.2.7
django-environ==0.11.2
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
django-storages==1.14.2
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-allauth==0.58.2
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-mptt==0.15.0
django-taggit==4.0.0
django-countries==7.5.1
django-phonenumber-field==7.2.0
phonenumbers==8.13.24

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache & Sessions
redis==5.0.1
hiredis==2.2.3

# Async Support
uvicorn==0.24.0
gunicorn==21.2.0
whitenoise==6.6.0

# Task Queue
celery==5.3.4
flower==2.0.1

# Authentication & Security
PyJWT==2.8.0
argon2-cffi==23.1.0
oauthlib==3.2.2
python-jose==3.3.0
cryptography==41.0.5

# Storage & Media
Pillow==10.1.0
boto3==1.29.3
python-magic==0.4.27

# Search
elasticsearch-dsl==7.4.1
django-elasticsearch-dsl==7.3.0

# Payment Processing
stripe==7.6.0
paypalrestsdk==1.13.1

# Email
sendgrid==6.10.0
django-anymail==10.1

# Performance & Optimization
django-cacheops==7.0.2
django-silk==5.0.4
django-querycount==0.8.3
django-optimized-image==0.3.0

# Monitoring & Logging
sentry-sdk==1.32.0
python-json-logger==2.0.7
django-request-logging==0.7.5
django-prometheus==2.3.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.0.3
coverage==7.3.2

# Development Tools
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1
pre-commit==3.5.0
django-stubs==4.2.6
djangorestframework-stubs==3.14.4

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
drf-spectacular==0.26.5

# API
drf-nested-routers==0.93.4
drf-yasg==1.21.7
django-filter==23.3
django-guardian==2.4.0

# Internationalization
Babel==2.13.1
django-rosetta==0.9.9
django-parler==2.3

# SEO
django-robots==5.0
django-meta==2.1.0
django-sitemap-generate==0.4.0

# Content Management
django-ckeditor==6.7.0
django-filebrowser==4.0.2
django-grappelli==3.0.7

# Social Media Integration
python-social-auth==0.3.6
social-auth-app-django==5.4.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
urllib3==2.0.7
python-slugify==8.0.1
Markdown==3.5.1
bleach==6.1.0
beautifulsoup4==4.12.2
html5lib==1.1
pycountry==22.3.5
phonenumbers==8.13.24
python-magic==0.4.27
python-dotenv==1.0.0

# Performance & Caching
django-cachalot==2.6.1
django-cache-memoize==0.1.10
django-cacheops==7.0.2

# Background Tasks
django-rq==2.8.1
rq==1.15.1
django-background-tasks==1.2.5

# Deployment
gunicorn==21.2.0
supervisor==4.2.5
nginx==1.24.0

# Version Control
gitpython==3.1.40

# Code Quality
bandit==1.7.5
safety==2.3.5
pylint==3.0.2
pylint-django==2.5.5
